requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
jinja2
python-docx
scikit-learn
python-multipart
pyyaml
